Avoiding ambiguity between two install functions (#31)
Currently there are two functions to install a community -
Install-TelligentCommunity and Install-TelligentInstance.  The latter is
specific to the LocalInstanceModule, which wraps the former with
conventions for a local installation.  However to someone just using help
to find hte function, the seperation between these two is unclear.

For now I made the base function internal and use some hackery to invoke
it from the Install-TelligentInstance function.  I'm not keen on this, but
can't come up with a better idea right now.

Fixes #21